multimodal-rag/requirements.txt
# FILE: requirements.txt
# Multimodal RAG Dependencies

# =============================================================================
# Core Framework
# =============================================================================
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pydantic>=2.0.0

# =============================================================================
# Document Processing
# =============================================================================
# PDF processing
pymupdf==1.24.0
pymupdf4llm==0.0.10

# Word documents
python-docx==1.1.0
docx2txt==0.8

# Excel/CSV
openpyxl==3.1.2
pandas>=2.0.0

# PowerPoint
python-pptx==0.6.23

# Images
Pillow>=10.0.0

# OCR (optional - requires tesseract system package)
pytesseract==0.3.10

# File type detection
python-magic==0.4.27

# =============================================================================
# Embeddings & ML
# =============================================================================
sentence-transformers==2.7.0
torch>=2.0.0
numpy>=1.24.0

# =============================================================================
# Vector Database
# =============================================================================
opensearch-py==2.5.0

# =============================================================================
# LLM & LangChain
# =============================================================================
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-text-splitters>=0.2.0
openai>=1.0.0

# =============================================================================
# Utilities
# =============================================================================
huggingface-hub>=0.20.0
tqdm>=4.66.0

# =============================================================================
# Testing
# =============================================================================
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
